--- routes/sweepapi_streams.js
@@ -1,10 +1,0 @@
-// require axios for promise based architecture 
-const axios = require('axios');
-
-// set sweepapi base url
-const sweepAPI = require('./sweepapi_auth');
-
-module.exports = function(app) {
-
-
-};
