+++ views/homepage.ejs
@@ -91,7 +91,7 @@
                   <td>placeholder</td>
                   <td>
                           
-                          <a href="/deleteDirectory" class="btn btn-sm btn-outline-primary" id="deleteParameterBtn" >Delete</a>
+                    <a href="/deleteDirectory" class="btn btn-sm btn-outline-primary" id="deleteParameterBtn" onclick="save" >Delete</a>
                   </td>
                 </tr>
               </form>
